refactor the client dashboard
